Change children in Nodes to be concrete instances
Because a global variable is used to know the size of each Point array,
it is now necessary to set g_vector_dimensions, e.g. in tests when using
the Point type. Otherwise a hard fail will be thrown in Point copy
constructor

The find_nearest_leaf function in pre-processing.cpp was re-implemented,
because the old implementation became unstable due the overall change to
concrete Nodes in children lists.